Option for serving domain from nfdhcpd to clients
On some occasions the clients do not send an fqdn as hostname,
so another way to send a domain is needed.

For that, a new optional config directive is introduced, called
"domain", as a way to hardcode the domain we serve.

If this directive is not defined, the traditional
'find domain through hostname' technique is used.